Changed type from int to byte; added plus, OR, AND, NOT and equals operations

+++ sc/sum1.c
@@ -1,4 +1,4 @@
-int a;
+byte a;
 void main() {
 	a = 1+2;
 }
